fix: Submission Review n°2: ETL
For the user_table_insert you have put DO NOTHING when there is a conflict. But would we like to ignore it? The level field is used for free and paid services for a user. Consider someone subscribes by paying a fee who happened to be a free user before. In that scenario, we would like to update the level of the user to paid. The current implementation will ignore the update, which is not desired. You can do that by ON CONFLICT (user_id) DO UPDATE SET level=EXCLUDED.level